+++ package.json
@@ -99,7 +99,7 @@
 		"react-native-tab-view": "^4.1.3",
 		"react-native-web": "^0.20.0",
 		"react-native-webview": "13.13.5",
-		"reanimated-composer": "^0.1.1",
+		"reanimated-composer": "^0.1.3",
 		"socket.io-client": "^4.8.1",
 		"tailwind-merge": "^3.2.0",
 		"tailwindcss": "^3.4.17",
