+++ src/features/signup/hooks/use-signup-progress.tsx
@@ -17,52 +17,6 @@ export type SignupForm = {
   appleId?: string;
 };
 
-type Agreement = {
-  id: string;
-  label: string;
-  link?: string;
-  required: boolean;
-  checked: boolean;
-};
-
-const AGREEMENTS: Agreement[] = [
-  {
-    id: "privacy",
-    label: "(필수) 개인정보 수집 및 이용 동의",
-    link: "https://ruby-composer-6d2.notion.site/1cd1bbec5ba180a3a4bbdf9301683145",
-    required: true,
-    checked: false,
-  },
-  {
-    id: "terms",
-    label: "(필수) 서비스 이용약관 동의",
-    link: "https://ruby-composer-6d2.notion.site/1cd1bbec5ba1805dbafbc9426a0aaa80",
-    required: true,
-    checked: false,
-  },
-  {
-    id: "location",
-    label: "(필수) 개인정보 처리방침 동의",
-    link: "https://ruby-composer-6d2.notion.site/1cd1bbec5ba180a3a4bbdf9301683145?pvs=73",
-    required: true,
-    checked: false,
-  },
-  {
-    id: "sensitive",
-    label: "(필수) 민감정보 이용 동의",
-    link: "https://www.notion.so/1cd1bbec5ba180ae800ff36c46285274",
-    required: true,
-    checked: false,
-  },
-  {
-    id: "marketing",
-    label: "(선택) 마케팅 수신 동의",
-    link: "https://www.notion.so/1cd1bbec5ba1800daa29fd7a8d01b7c9",
-    required: false,
-    checked: false,
-  },
-];
-
 type StoreProps = {
   progress: number;
   step: SignupSteps;
@@ -74,20 +28,18 @@ type StoreProps = {
   univTitle: string;
   updateUnivTitle: (area: string) => void;
   updateForm: (form: Partial<SignupForm>) => void;
-  agreements: Agreement[];
-  updateAgreements: (agreements: Agreement[]) => void;
+
   clear: () => void;
-  smsComplete: boolean;
-  completeSms: () => void;
+
   updateShowHeader: (bool: boolean) => void;
 };
 
 export enum SignupSteps {
-  AREA = 1,
-  UNIVERSITY = 2,
-  UNIVERSITY_DETAIL = 3,
-  INSTAGRAM = 4,
-  PROFILE_IMAGE = 5,
+  // AREA = 1,
+  UNIVERSITY = 1,
+  UNIVERSITY_DETAIL = 2,
+  INSTAGRAM = 3,
+  PROFILE_IMAGE = 4,
 }
 
 const phaseCount = Object.keys(SignupSteps).length / 2;
@@ -95,7 +47,7 @@ const phaseCount = Object.keys(SignupSteps).length / 2;
 const useSignupProgress = create<StoreProps>((set) => ({
   progress: 1 / phaseCount,
 
-  step: SignupSteps.AREA,
+  step: SignupSteps.UNIVERSITY,
   updateStep: (step) => {
     const isLast = step === phaseCount;
     return set({
@@ -116,9 +68,8 @@ const useSignupProgress = create<StoreProps>((set) => ({
   clear: () =>
     set({
       form: {},
-      step: SignupSteps.AREA,
-      smsComplete: false,
-      agreements: AGREEMENTS,
+      step: SignupSteps.UNIVERSITY,
+
       univTitle: "",
     }),
 
@@ -127,12 +78,7 @@ const useSignupProgress = create<StoreProps>((set) => ({
   },
   showHeader: false,
   univTitle: "",
-  updateUnivTitle: (area: string) => set({ univTitle: area }),
-  agreements: AGREEMENTS,
-  updateAgreements: (agreements) => set({ agreements }),
-
-  smsComplete: false,
-  completeSms: () => set({ smsComplete: true }),
+  updateUnivTitle: (univTitle: string) => set({ univTitle: univTitle }),
 }));
 
 export default useSignupProgress;
