+++ src/features/signup/queries/use-universities.tsx
@@ -7,7 +7,6 @@ import {
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { useGlobalSearchParams } from "expo-router";
 import { getUniversitiesByRegion } from "../apis";
-import { useSignupProgress } from "../hooks";
 import { getAllRegionList, getRegionsByRegionCode } from "../lib";
 
 export default function useUniversities() {
